+++ design-system/dropdown-menu/examples/constellation/dropdown-item-multiline.tsx
@@ -5,7 +5,10 @@ import DropdownMenu, { DropdownItem, DropdownItemGroup } from '../../src';
 const DropdownItemMultilineExample = () => (
   <DropdownMenu triggerType="button">
     <DropdownItemGroup>
-      <DropdownItem shouldAllowMulti>Sydney is a beautiful city</DropdownItem>
+      <DropdownItem shouldAllowMultiline>
+        Sydney is capital of New South Wales and one of Australia's largest
+        cities.
+      </DropdownItem>
       <DropdownItem>Brisbane</DropdownItem>
     </DropdownItemGroup>
   </DropdownMenu>