+++ design-system/dropdown-menu/examples/constellation/dropdown-menu-should-flip.tsx
@@ -5,7 +5,8 @@ import DropdownMenu, { DropdownItem, DropdownItemGroup } from '../../src';
 const DropdownMenuShouldFlipExample = () => (
   <DropdownMenu triggerType="button" shouldFlip>
     <DropdownItemGroup>
-      <DropdownItem>Sydney</DropdownItem>
+      <DropdownItem id="sydney">Sydney</DropdownItem>
+      <DropdownItem id="melbourne">Melbourne</DropdownItem>
     </DropdownItemGroup>
   </DropdownMenu>
 );