+++ design-system/dropdown-menu/examples/constellation/dropdown-menu-stateless-boundaries.tsx
@@ -17,7 +17,7 @@ const StatelessMenuBoundriesExample = () => {
         onOpenChange={(attrs: OnOpenChangeArgs) => {
           setIsOpen(attrs.isOpen);
         }}
-        trigger="Filter City"
+        trigger="Filter city"
         triggerType="button"
         boundariesElement="viewport"
       >