+++ design-system/dropdown-menu/examples/constellation/dropdown-menu-stateless-on-open-change.tsx
@@ -17,7 +17,7 @@ const StatelessMenuOnOpenChangeExample = () => {
         onOpenChange={(attrs: OnOpenChangeArgs) => {
           setIsOpen(attrs.isOpen);
         }}
-        trigger="Filter City"
+        trigger="Filter city"
         triggerType="button"
       >
         <DropdownItemGroupRadio id="cities">