+++ design-system/menu/examples/link-item.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-import { LinkItem } from '../src';
+import { LinkItem, LinkItemProps } from '../src';
 
 import koala from './icons/koala.png';
 
@@ -8,28 +8,64 @@ const ImgIcon = ({ src, alt }: { src: string; alt: string }) => (
   <img src={src} height={24} width={24} alt={alt} style={{ borderRadius: 3 }} />
 );
 
-export default () => (
-  <div onClick={e => e.preventDefault()}>
-    <LinkItem href="link-item">Customer Feedback</LinkItem>
-    <LinkItem isSelected href="link-item">
-      Customer Feedback
-    </LinkItem>
-    <LinkItem isDisabled href="link-item">
-      Customer Feedback
-    </LinkItem>
+const useLinkItemComputedProps = ({
+  initialSelectedHref,
+}: {
+  initialSelectedHref?: string;
+}) => {
+  const [currentHref, setCurrentHref] = useState<string | undefined>(
+    initialSelectedHref,
+  );
 
-    <LinkItem href="link-item" description="Classic service desk">
-      Customer Feedback
-    </LinkItem>
-    <LinkItem
-      href="link-item"
-      iconBefore={<ImgIcon src={koala} alt={'A koala'} />}
-      description="Classic service desk"
-    >
-      Customer Feedback
-    </LinkItem>
-    <LinkItem href="https://atlassian.design" testId="link-item">
-      Atlassian Design
-    </LinkItem>
-  </div>
-);
+  const getComputedProps = ({ href, ...restProps }: LinkItemProps) => ({
+    href,
+    ...restProps,
+    isSelected: currentHref === href,
+    onClick: () => {
+      setCurrentHref(href);
+    },
+  });
+
+  return {
+    getComputedProps,
+  };
+};
+
+export default () => {
+  const { getComputedProps } = useLinkItemComputedProps({
+    initialSelectedHref: '#link-item2',
+  });
+
+  return (
+    <div onClick={e => e.preventDefault()}>
+      <LinkItem {...getComputedProps({ href: '#link-item1' })}>
+        Customer Feedback
+      </LinkItem>
+      <LinkItem {...getComputedProps({ href: '#link-item2' })}>
+        Customer Feedback
+      </LinkItem>
+      <LinkItem {...getComputedProps({ href: '#link-item3' })} isDisabled>
+        Customer Feedback
+      </LinkItem>
+      <LinkItem
+        {...getComputedProps({ href: '#link-item4' })}
+        description="Classic service desk"
+      >
+        Customer Feedback
+      </LinkItem>
+      <LinkItem
+        {...getComputedProps({ href: '#link-item5' })}
+        iconBefore={<ImgIcon src={koala} alt={'A koala'} />}
+        description="Classic service desk"
+      >
+        Customer Feedback
+      </LinkItem>
+      <LinkItem
+        {...getComputedProps({ href: 'https://atlassian.design' })}
+        testId="link-item"
+      >
+        Atlassian Design
+      </LinkItem>
+    </div>
+  );
+};