+++ design-system/select/examples/constellation/select-country.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { CountrySelect } from '../../src';
+
+const CountrySelectExample = () => <CountrySelect placeholder="Country" />;
+
+export default CountrySelectExample;